[retrieve] retrieve profiles only + retrieve specific files

+++ src/sftx-retrieve.js
@@ -12,7 +12,7 @@ const stripEmptyTranslations = require('./prepare/strip-empty-translations')
 const stripUselessFlsInPermissionSets = require('./prepare/strip-useless-fls-in-permission-sets')
 const stripPartnerRoles = require('./prepare/strip-partner-roles')
 const fixProfiles = require('./prepare/fix-profiles')
-const { getMembersOf } = require('./utils/package-utils')
+const { getMembersOf, getProfileOnlyPackage, getPackageXml } = require('./utils/package-utils')
 const multimatch = require('multimatch')
 require('./error-handling')()
 
@@ -20,6 +20,8 @@ program
   .option('-u, --username <username>', 'Username')
   .option('-p, --password <password>', 'Password + Token')
   .option('-s, --sandbox', 'Use sandbox login endpoint')
+  .option('-P, --profile-only', 'Retrieve profiles only')
+  .option('--files <files>', 'Retrieve specific files')
   .parse(process.argv)
 
 if (!program.username || !program.password) {
@@ -32,6 +34,7 @@ if (!fs.existsSync(configPath)) throw Error('Missing configuration file .sftx.js
 const config = require(configPath)
 
 ;(async () => {
+  console.time('running time')
   log(chalk.green('SFTX V1.0'))
   log(chalk.yellow(`(1/4) Logging in salesforce as ${program.username}...`))
   const sfdcConnector = await Sfdc.newInstance({
@@ -40,27 +43,44 @@ const config = require(configPath)
     isSandbox: !!program.sandbox
   })
   log(chalk.green(`Logged in!`))
-  const srcFolder = path.resolve(process.cwd(), 'src')
-  const packageXmlPath = path.resolve(srcFolder, 'package.xml')
   log(chalk.yellow(`(2/4) Retrieving metadata...`))
-  const retrieveJob = await sfdcConnector.retrieveMetadata(packageXmlPath, config.profiles && config.profiles.addExtraApplications)
+  if (program.profileOnly) log(chalk.yellow(`--profile-only=true. Retrieving profiles only...`))
+  const specificFiles = (program.files && program.files.split(',').map(x => x.trim())) || []
+  if (specificFiles.length) log(chalk.yellow(`--files specified. Retrieving only specific files...`))
+  const pkgJson = (await (
+    program.profileOnly
+      ? getProfileOnlyPackage()
+      : getPackageXml({
+        specificFiles,
+        sfdcConnector
+      })
+  )).Package
+  if (specificFiles.length) log(chalk.yellow(`delta package generated`))
+
+  const retrieveJob = await sfdcConnector.retrieveMetadata(pkgJson, config.profiles && config.profiles.addExtraApplications)
   const retrieveResult = await sfdcConnector.pollRetrieveMetadataStatus(retrieveJob.id)
   log(chalk.green(`Retrieve completed!`))
   log(chalk.yellow(`(3/4) Unzipping...`))
   const zipBuffer = await b64.decode(retrieveResult.zipFile)
-  await decompress(zipBuffer, srcFolder, {
-    filter: f => !/package\.xml$/.test(f.path)
+  await decompress(zipBuffer, path.resolve(process.cwd(), 'src'), {
+    filter: f => program.profileOnly ? f.path.endsWith('.profile') : !/package\.xml$/.test(f.path)
   })
   log(chalk.green(`Unzipped!`))
   log(chalk.yellow(`(4/4) Applying patches...`))
+  const patchProfiles = pkgJson.types.some(x => x.name[0] === 'Profile')
+  const patchTranslations = pkgJson.types.some(x => x.name[0] === 'CustomObjectTranslation')
+  const patchPermissionSet = pkgJson.types.some(x => x.name[0] === 'PermissionSet')
+  const patchPartnerRoles = pkgJson.types.some(x => x.name[0] === 'Role')
+
   await Promise.all([
-    stripEmptyTranslations(config),
-    stripUselessFlsInPermissionSets(config),
-    fixProfiles(config, sfdcConnector)
+    patchTranslations ? stripEmptyTranslations(config) : Promise.resolve(),
+    patchPermissionSet ? stripUselessFlsInPermissionSets(config) : Promise.resolve(),
+    patchProfiles ? fixProfiles(config, sfdcConnector) : Promise.resolve(),
+    patchPartnerRoles ? Promise.resolve(stripPartnerRoles(config)) : Promise.resolve()
   ])
-  stripPartnerRoles(config)
+
   const APPS_PATH = path.resolve(process.cwd(), 'src', 'applications')
-  if (fs.existsSync(APPS_PATH)) {
+  if (fs.existsSync(APPS_PATH) && pkgJson.types.some(x => x.name[0] === 'Profile')) {
     const versionedApps = (await getMembersOf('CustomApplication')).map(x => x + '.app')
     if (!versionedApps.length) fs.removeSync(APPS_PATH)
     else {
@@ -70,4 +90,5 @@ const config = require(configPath)
     }
   }
   log(chalk.green(`Patches applied!`))
+  console.timeEnd('running time')
 })()
